textrender: Slightly cleaner form of touch interaction
Instead of having QML intercept presses and send them to textrender for
gesture processing, have textrender intercept them from the Keys by
making them children of textrender. Now that we have a contentItem for
textrender, this doesn't affect anything visually, so all is well.

This will also allow us to introduce more interesting gesture processing
through touch events in the future if we so desire.