textrender: Don't re-render on change instantly
We get a flood of changes from e.g. terminal constantly. Re-polishing
every frame takes a toll on heavy updates, so debounce updates into
something a bit less frequent. This decreases the amount of time to
print a huge amount of text a fair bit.

Note that we may need to tweak this some to make resizes still
synchronous, but that's future work.