Fix keyboard opacity, or at least make it better
The element setup was changed after Fingerterm fork. Earlier keyboard
was bottom element and opacity changing terminal was on top of it.

Now terminal is on bottom and keyobard is over it.

Changed opacities so the end result is roughly the same as earlier.
Some little difference can be seen if there's text beneath the keyboard.