Split SGR handling out to a separate file ...
... and make it stateless, acting on input parameters, rather than the struct.
This way, we can write tests for it: which we also do in this
commit (using catch2).

One functional change: the code will now drop any SGR which is invalid,
rather than continuing to search it for changes. I deem this to be better than
dealing with malformed input that most likely won't make sense, but if
it causes problems, it may need to be revisited somehow.